refactor(billing): extract schedule context helper
